requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
docker==7.0.0
idna==3.6
packaging==23.2
psutil==5.9.8
pyTelegramBotAPI==4.9.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.2.0